Pre-generate Bedrock biome definitions
