+++ tools/pumpkin-codegen/src/main.rs
@@ -27,6 +27,7 @@ use std::{
 
 mod advancement;
 mod attributes;
+mod bedrock_biome;
 mod bedrock_creative;
 mod biome;
 mod bitsets;
@@ -97,6 +98,7 @@ pub fn main() {
 
     let mut build_functions: Vec<(BuilderFn, &str)> = vec![
         (advancement::build, "advancement.rs"),
+        (bedrock_biome::build, "bedrock_biome.rs"),
         (bedrock_creative::build, "bedrock_creative.rs"),
         (packet::build, "packet.rs"),
         (screen::build, "screen.rs"),
